chore: move vote elements to separate components
